chore : modify validation path & alias
p.s barrel 파일에서 export 하지 않는 파일 삭제
