--- apps/client/src/app/auth/_constants/validation/index.ts
@@ -1,3 +1,0 @@
-export { signUpValidationMsg } from "@auth/_constants/validation/validationMessages";
-export { formValidationRegex } from "@auth/_constants/validation/validationRegex";
-export { signUpBuyerSchema, signUpSellerSchema } from "@auth/_constants/validation/validationSchemas";
