+++ apps/client/src/app/auth/_constants/validation/signup/index.ts
@@ -1,0 +1,1 @@
+export { signUpBuyerSchema, signUpSellerSchema } from "~/app/auth/_constants/validation/signup/validationSchemas";
