+++ apps/client/src/app/auth/_constants/validation/signup/validationMsg.ts
@@ -1,4 +1,4 @@
-export const signUpValidationMsg = {
+export const signUpErrMsg = {
   email: {
     required: "이메일을 입력해주세요.",
     pattern: "올바른 이메일 형식으로 입력해주세요.",
